requirements.txt
Django==3.1.5
django-npm==1.0.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
gunicorn==20.0.4
django-heroku==0.3.1
whitenoise==5.2.0